requirements.txt
streamlit
pandas
openpyxl
matplotlib
seaborn
google-generativeai
python-dotenv